+++ src/Data/Bytes/Parser/Latin.hs
@@ -29,6 +29,10 @@ module Data.Bytes.Parser.Latin
   , char5
   , char6
   , char7
+  , char8
+  , char9
+  , char10
+  , char11
     -- ** Try
   , trySatisfy
   , trySatisfyThen
@@ -213,6 +217,79 @@ char7 e !c0 !c1 !c2 !c3 !c4 !c5 !c6 = uneffectful $ \chunk ->
      , indexLatinCharArray (array chunk) (offset chunk + 5) == c5
      , indexLatinCharArray (array chunk) (offset chunk + 6) == c6
          -> InternalSuccess () (offset chunk + 7) (length chunk - 7)
+     | otherwise -> InternalFailure e
+
+-- | Consume eight characters, failing if they do
+-- not match the expected values.
+char8 :: e -> Char -> Char -> Char -> Char -> Char -> Char -> Char -> Char -> Parser e s ()
+char8 e !c0 !c1 !c2 !c3 !c4 !c5 !c6 !c7 = uneffectful $ \chunk ->
+  if | length chunk > 7
+     , indexLatinCharArray (array chunk) (offset chunk) == c0
+     , indexLatinCharArray (array chunk) (offset chunk + 1) == c1
+     , indexLatinCharArray (array chunk) (offset chunk + 2) == c2
+     , indexLatinCharArray (array chunk) (offset chunk + 3) == c3
+     , indexLatinCharArray (array chunk) (offset chunk + 4) == c4
+     , indexLatinCharArray (array chunk) (offset chunk + 5) == c5
+     , indexLatinCharArray (array chunk) (offset chunk + 6) == c6
+     , indexLatinCharArray (array chunk) (offset chunk + 7) == c7
+         -> InternalSuccess () (offset chunk + 8) (length chunk - 8)
+     | otherwise -> InternalFailure e
+
+-- | Consume nine characters, failing if they do
+-- not match the expected values.
+char9 :: e -> Char -> Char -> Char -> Char
+  -> Char -> Char -> Char -> Char -> Char -> Parser e s ()
+char9 e !c0 !c1 !c2 !c3 !c4 !c5 !c6 !c7 !c8 = uneffectful $ \chunk ->
+  if | length chunk > 8
+     , indexLatinCharArray (array chunk) (offset chunk) == c0
+     , indexLatinCharArray (array chunk) (offset chunk + 1) == c1
+     , indexLatinCharArray (array chunk) (offset chunk + 2) == c2
+     , indexLatinCharArray (array chunk) (offset chunk + 3) == c3
+     , indexLatinCharArray (array chunk) (offset chunk + 4) == c4
+     , indexLatinCharArray (array chunk) (offset chunk + 5) == c5
+     , indexLatinCharArray (array chunk) (offset chunk + 6) == c6
+     , indexLatinCharArray (array chunk) (offset chunk + 7) == c7
+     , indexLatinCharArray (array chunk) (offset chunk + 8) == c8
+         -> InternalSuccess () (offset chunk + 9) (length chunk - 9)
+     | otherwise -> InternalFailure e
+
+-- | Consume ten characters, failing if they do
+-- not match the expected values.
+char10 :: e -> Char -> Char -> Char -> Char -> Char
+  -> Char -> Char -> Char -> Char -> Char -> Parser e s ()
+char10 e !c0 !c1 !c2 !c3 !c4 !c5 !c6 !c7 !c8 !c9 = uneffectful $ \chunk ->
+  if | length chunk > 9
+     , indexLatinCharArray (array chunk) (offset chunk) == c0
+     , indexLatinCharArray (array chunk) (offset chunk + 1) == c1
+     , indexLatinCharArray (array chunk) (offset chunk + 2) == c2
+     , indexLatinCharArray (array chunk) (offset chunk + 3) == c3
+     , indexLatinCharArray (array chunk) (offset chunk + 4) == c4
+     , indexLatinCharArray (array chunk) (offset chunk + 5) == c5
+     , indexLatinCharArray (array chunk) (offset chunk + 6) == c6
+     , indexLatinCharArray (array chunk) (offset chunk + 7) == c7
+     , indexLatinCharArray (array chunk) (offset chunk + 8) == c8
+     , indexLatinCharArray (array chunk) (offset chunk + 9) == c9
+         -> InternalSuccess () (offset chunk + 10) (length chunk - 10)
+     | otherwise -> InternalFailure e
+
+-- | Consume ten characters, failing if they do
+-- not match the expected values.
+char11 :: e -> Char -> Char -> Char -> Char -> Char -> Char
+  -> Char -> Char -> Char -> Char -> Char -> Parser e s ()
+char11 e !c0 !c1 !c2 !c3 !c4 !c5 !c6 !c7 !c8 !c9 !c10 = uneffectful $ \chunk ->
+  if | length chunk > 10
+     , indexLatinCharArray (array chunk) (offset chunk) == c0
+     , indexLatinCharArray (array chunk) (offset chunk + 1) == c1
+     , indexLatinCharArray (array chunk) (offset chunk + 2) == c2
+     , indexLatinCharArray (array chunk) (offset chunk + 3) == c3
+     , indexLatinCharArray (array chunk) (offset chunk + 4) == c4
+     , indexLatinCharArray (array chunk) (offset chunk + 5) == c5
+     , indexLatinCharArray (array chunk) (offset chunk + 6) == c6
+     , indexLatinCharArray (array chunk) (offset chunk + 7) == c7
+     , indexLatinCharArray (array chunk) (offset chunk + 8) == c8
+     , indexLatinCharArray (array chunk) (offset chunk + 9) == c9
+     , indexLatinCharArray (array chunk) (offset chunk + 10) == c10
+         -> InternalSuccess () (offset chunk + 11) (length chunk - 11)
      | otherwise -> InternalFailure e
 
 -- | Consumes and returns the next character in the input.